feat: Refactor InfoMessage and introduce GoalMessage component
- Removed the old InfoMessage component and consolidated its functionality into a new Message component.
- Added GoalMessage component to provide a distinct message type with a different icon.
- Updated common index file to export both InfoMessage and GoalMessage from the new Message file.
- Refactored LearningSection to import the updated components for improved message handling.
