+++ src/components/common/index.ts
@@ -2,4 +2,4 @@ export { default as Title } from './Title';
 export { default as Circle } from './Circle';
 export { default as DescriptionSection } from './DescriptionSection';
 export { default as CardLayout } from './CardLayout';
-export { default as InfoMessage } from './InfoMessage';
+export { InfoMessage, GoalMessage } from './Message';
